Removed hard coded passwords. Follow the test framework to load credentials from test.properties
