+++ ojdbc-provider-jackson-oson/src/test/java/oracle/jdbc/provider/oson/test/AllTypesTest.java
@@ -1,4 +1,42 @@
-package oson.test;
+/*
+ ** Copyright (c) 2024 Oracle and/or its affiliates.
+ **
+ ** The Universal Permissive License (UPL), Version 1.0
+ **
+ ** Subject to the condition set forth below, permission is hereby granted to any
+ ** person obtaining a copy of this software, associated documentation and/or data
+ ** (collectively the "Software"), free of charge and under any and all copyright
+ ** rights in the Software, and any and all patent rights owned or freely
+ ** licensable by each licensor hereunder covering either (i) the unmodified
+ ** Software as contributed to or provided by such licensor, or (ii) the Larger
+ ** Works (as defined below), to deal in both
+ **
+ ** (a) the Software, and
+ ** (b) any piece of software and/or hardware listed in the lrgrwrks.txt file if
+ ** one is included with the Software (each a "Larger Work" to which the Software
+ ** is contributed by such licensors),
+ **
+ ** without restriction, including without limitation the rights to copy, create
+ ** derivative works of, display, perform, and distribute the Software and make,
+ ** use, sell, offer for sale, import, export, have made, and have sold the
+ ** Software and the Larger Work(s), and to sublicense the foregoing rights on
+ ** either these or other terms.
+ **
+ ** This license is subject to the following condition:
+ ** The above copyright notice and either this complete permission notice or at
+ ** a minimum a reference to the UPL must be included in all copies or
+ ** substantial portions of the Software.
+ **
+ ** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ ** IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ ** FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ ** AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ ** LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ ** OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ ** SOFTWARE.
+ */
+
+package oracle.jdbc.provider.oson.test;
 
 import com.fasterxml.jackson.core.exc.StreamWriteException;
 import com.fasterxml.jackson.databind.DatabindException;
@@ -6,10 +44,12 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import oracle.jdbc.OracleType;
 import oracle.jdbc.datasource.impl.OracleDataSource;
+import oracle.jdbc.provider.TestProperties;
 import oracle.jdbc.provider.oson.OsonFactory;
 import oracle.jdbc.provider.oson.OsonGenerator;
 import oracle.jdbc.provider.oson.OsonParser;
-import oson.model.AllOracleTypes;
+import oracle.jdbc.provider.oson.OsonTestProperty;
+import oracle.jdbc.provider.oson.model.AllOracleTypes;
 import oracle.sql.json.OracleJsonDatum;
 import oracle.sql.json.OracleJsonObject;
 import oracle.sql.json.OracleJsonValue;
@@ -56,13 +96,17 @@ public class AllTypesTest {
 	@BeforeAll
 	public void setup() {
 		try {
+			String url = TestProperties.getOrAbort(OsonTestProperty.JACKSON_OSON_URL);
+			String userName = TestProperties.getOrAbort(OsonTestProperty.JACKSON_OSON_USERNAME);
+			String password = TestProperties.getOrAbort(OsonTestProperty.JACKSON_OSON_PASSWORD);
+
 			om.findAndRegisterModules();
 			osonGen = (OsonGenerator) osonFactory.createGenerator(out);
 			
 			OracleDataSource ods = new OracleDataSource();
-			ods.setURL("jdbc:oracle:thin:@(DESCRIPTION=(ADDRESS=(PROTOCOL=TCP)(HOST=phoenix93464.dev3sub2phx.databasede3phx.oraclevcn.com)(PORT=5521))(CONNECT_DATA=(SERVICE_NAME=cdb1_pdb1.regress.rdbms.dev.us.oracle.com)))");
-			ods.setUser("system");
-			ods.setPassword("manager");
+			ods.setURL(url);
+			ods.setUser(userName);
+			ods.setPassword(password);
 			conn = ods.getConnection();
 			
 			//setup db tables
@@ -80,7 +124,7 @@ public class AllTypesTest {
 	@Test
 	@Order(1)
 	public void convertToOson() throws StreamWriteException, DatabindException, IOException {
-		
+		Assumptions.assumeTrue(conn != null);
 		om.writeValue(osonGen, allTypes);
 		osonGen.close();
 		osonTocheck = out.toByteArray();
@@ -92,6 +136,7 @@ public class AllTypesTest {
 	@Test
 	@Order(2)
 	public void convertFromOson() throws Exception {
+		Assumptions.assumeTrue(conn != null);
 		try (OsonParser parser = (OsonParser) osonFactory.createParser(osonTocheck) ) {
 			
 			AllOracleTypes ne = om.readValue(parser, AllOracleTypes.class);
@@ -103,6 +148,7 @@ public class AllTypesTest {
 	@Test
 	@Order(3)
 	public void insertIntoDB() throws Exception {
+		Assumptions.assumeTrue(conn != null);
 		try (OsonParser parser = (OsonParser) osonFactory.createParser(osonTocheck) ) {
 			
 			// insert into db
@@ -118,8 +164,8 @@ public class AllTypesTest {
 	@Test
 	@Order(4)
 	public void retieveFromDatabase() throws Exception {
-    
-    //retrieve from db
+		Assumptions.assumeTrue(conn != null);
+        //retrieve from db
 		Statement stmt = conn.createStatement();
 		ResultSet rs = stmt.executeQuery("select c1, c2 from all_types_json order by c1");
 		while(rs.next()) {
