+++ ojdbc-provider-samples/example-configuration.properties
@@ -115,5 +115,5 @@ gcp_cloud_storage_properties=project=onyx-eye-426013-i5;bucket=fm-test-bucket-12
 
 # Jackson OSON provider
 jackson_oson_url=jdbc:oracle:thin:@tcp://phoenix127228.dev3sub3phx.databasede3phx.oraclevcn.com:5521/cdb1_pdb1.regress.rdbms.dev.us.oracle.com
-jackson_oson_username=scott
-jackson_oson_password=tiger
+jackson_oson_username=user_name
+jackson_oson_password=password
